+++ src/QuickMarcEditor/index.js
@@ -1,2 +1,4 @@
 export { default as QuickMarcEditor } from './QuickMarcEditor';
 export { default as QuickMarcEditorContainer } from './QuickMarcEditorContainer';
+export { default as QuickMarcDuplicateWrapper } from './QuickMarcDuplicateWrapper';
+export { default as QuickMarcEditWrapper } from './QuickMarcEditWrapper';